+++ vnext/ReactUWP/Utils/Helpers.cpp
@@ -4,9 +4,9 @@
 #include "pch.h"
 
 #include <Modules/NativeUIManager.h>
+#include <Utils/Helpers.h>
 #include <winrt/Windows.Foundation.Metadata.h>
 #include <winrt/Windows.UI.Xaml.Media.h>
-#include "Helpers.h"
 
 namespace winrt {
 using namespace Windows::UI::Xaml::Controls::Primitives;
