+++ vnext/ReactUWP/Utils/UwpScriptStore.cpp
@@ -1,11 +1,11 @@
 #include "pch.h"
 
+#include <Utils/UwpScriptStore.h>
 #include <winrt/Windows.Foundation.h>
 #include <winrt/Windows.Storage.FileProperties.h>
 #include <winrt/Windows.Storage.h>
 #include <future>
 #include "Unicode.h"
-#include "Utils/UwpScriptStore.h"
 
 namespace winrt {
 using namespace winrt::Windows::Foundation;
