+++ vnext/include/ReactUWP/Utils/CppWinrtLessExceptions.h
