+++ vnext/include/ReactUWP/Utils/Helpers.h
