+++ vnext/include/ReactUWP/Utils/LocalBundleReader.h
