+++ vnext/include/ReactUWP/Utils/PropertyHandlerUtils.h
