+++ vnext/include/ReactUWP/Utils/PropertyUtils.h
