+++ vnext/include/ReactUWP/Utils/ResourceBrushUtils.h
