+++ vnext/include/ReactUWP/Utils/StandardControlResourceKeyNames.h
