+++ vnext/include/ReactUWP/Utils/UwpPreparedScriptStore.h
