+++ vnext/include/ReactUWP/Utils/UwpScriptStore.h
